# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "A Social platform for the creative at heart",
    "version": "1.0.0",
    "title": "s tutela",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "ninjas-ah-backend.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/auth/facebook": {
      "get": {
        "tags": ["auth"],
        "summary": "Register/login a facebook user",
        "description": "Register/login a facebook user and fetch the user profile",
        "operationId": "Facebook",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User successfully authenticated"
          },
          "201": {
            "description": "Account Successfully created"
          },
          "409": {
            "description": "User alredy exist"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/auth/google": {
      "get": {
        "tags": ["auth"],
        "summary": "Register/login a google user",
        "description": "Register/login a google user and fetch the user profile",
        "operationId": "Google",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User successfully authenticated"
          },
          "201": {
            "description": "Account Successfully created"
          },
          "409": {
            "description": "User alredy exist"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/auth/twitter": {
      "get": {
        "tags": ["auth"],
        "summary": "Register/login a twitter user",
        "description": "Register/login a twitter user and fetch the user profile",
        "operationId": "Twitter",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User successfully authenticated"
          },
          "201": {
            "description": "Account Successfully created"
          },
          "409": {
            "description": "User alredy exist"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "tags": ["auth"],
        "summary": "Logs out current logged in user",
        "description": "Blacklist a token",
        "operationId": "LogoutUser",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User successfully logged out"
          },
          "401": {
            "description": "User not ized to blacklist a given token"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/articles/{slug}": {
      "get": {
        "tags": ["articles"],
        "summary": "Get articles",
        "description": "Get articles",
        "operationId": "GetArticle",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Article successfully fetched"
          },
          "404": {
            "description": "Article not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "tags": ["articles"],
        "summary": "Update an article",
        "description": "Update an article",
        "operationId": "UpdateArticle",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Information to be sent to create an article",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["title", "description", "body"],
              "properties": {
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "body": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Article successfully updated"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not ized to update an article"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": ["articles"],
        "summary": "Delete article",
        "description": "Delete article",
        "operationId": "DeleteArticle",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Article successfully deleted"
          },
          "401": {
            "description": "Not ized to delete article"
          },
          "404": {
            "description": "Article not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "access-token",
      "in": "header"
    }
  },
...
}